Declare mixins first, move mixin to shared file
Mixins should be listed in class definition lists prior to the
superclass.
Since other components may also need mixins, they
have been moved to a new file.
