+++ elyra/metadata/runtime.py
@@ -15,6 +15,7 @@
 #
 
 from .metadata import Metadata, MetadataManager
+from .mixins import AppUtilMixin
 
 from traitlets.config.application import Application
 from jupyter_core.application import base_flags
@@ -29,21 +30,7 @@ class Runtime(Metadata):
     namespace = 'runtimes'
 
 
-class AppUtilMixin(object):
-
-    def _log_and_exit(self, msg, exit_status=1, display_help=False):
-        self.log.error(msg)
-        if display_help:
-            print()
-            self.print_help()
-        self.exit(exit_status)
-
-    def _confirm_required(self, name, value):
-        if value is None or len(value) == 0:
-            self._log_and_exit("'{}' is a required parameter.".format(name), display_help=True)
-
-
-class ListRuntimes(Application, AppUtilMixin):
+class ListRuntimes(AppUtilMixin, Application):
     version = __version__
     description = """List installed external runtime metadata."""
 
@@ -97,7 +84,7 @@ class ListRuntimes(Application, AppUtilMixin):
                                         invalid))
 
 
-class RemoveRuntime(Application, AppUtilMixin):
+class RemoveRuntime(AppUtilMixin, Application):
     version = __version__
     description = """Remove external runtime metadata."""
 
@@ -124,7 +111,7 @@ class RemoveRuntime(Application, AppUtilMixin):
         self._confirm_required("name", self.name)
 
 
-class Kfp(Application, AppUtilMixin):
+class Kfp(AppUtilMixin, Application):
     version = __version__
     description = """Install runtime metadata for Kubeflow pipelines."""
 
@@ -215,7 +202,7 @@ class Kfp(Application, AppUtilMixin):
         self._confirm_required("cos_bucket", self.cos_bucket)
 
 
-class Airflow(Application, AppUtilMixin):
+class Airflow(AppUtilMixin, Application):
     version = __version__
     description = """Install runtime metadata for Airflow pipelines."""
     flags = {}
